style: update style of item list in tool subpage
