+++ src/app/[locale]/tools/kieu-tools/page.tsx
@@ -1,6 +1,5 @@
-import { Link } from "@/i18n/routing";
-
 import BreadcrumbAndSearchBar from "@/components/layout/BreadcrumbAndSearchBar";
+import { ToolList } from "../ToolList";
 import { getTranslations } from "next-intl/server";
 
 import { Merriweather } from "next/font/google";
@@ -10,7 +9,15 @@ const merriweather = Merriweather({ weight: "300", subsets: ["vietnamese"] });
 
 const KieuTools = async ({ params: { locale } }) => {
   const t = await getTranslations();
-
+  const tools = [
+    {
+      href: "/tools/kieu-tools/character-frequency",
+      title: t("Tools.kieu-tools.tools.glossary.name"),
+      description: t("Tools.kieu-tools.tools.glossary.description"),
+      target: "",
+      rel: "",
+    },
+  ];
   return (
     <div className="flex flex-col items-center max-width">
       <div className="flex-col mb-20 w-full">
@@ -38,10 +45,10 @@ const KieuTools = async ({ params: { locale } }) => {
         <div className="mt-28">
           <Separator />
         </div>
-        <div className="mt-10 text-muted-foreground">To be implemented</div>
-        <Link href="/tools/kieu-tools/character-frequency">
-          <div className="mt-10 text-branding-black">Character Frequency</div>
-        </Link>
+
+        <div className="mt-10">
+          <ToolList tools={tools} />
+        </div>
       </div>
     </div>
   );
